backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-simple-history==3.8.0
djangorestframework==3.16.0
drf-yasg==1.21.10
ecdsa==0.19.1
inflection==0.5.1
mysqlclient==2.2.7
packaging==25.0
pyasn1==0.4.8
python-dotenv==1.1.0
python-jose==3.4.0
pytz==2025.2
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
